Removed redux18in
