+++ lib/withI18next.js
@@ -11,8 +11,10 @@ export const withI18next = (namespaces = ["common"]) => ComposedComponent => {
       ? await ComposedComponent.getInitialProps(ctx)
       : {};
 
-    const i18nInitialProps =
-      ctx.req && !process.browser ? getInitialProps(ctx.req, namespaces) : {};
+    let i18nInitialProps =
+      ctx.req && !process.browser
+        ? getInitialProps(ctx.req, namespaces, ctx.store)
+        : {};
 
     return {
       ...composedInitialProps,
