+++ server.js
@@ -18,7 +18,6 @@ const { i18nInstance } = require("./i18n");
 const deploy = require("./utils/deploy_notification");
 
 const airTable = require("./utils/airtable_es2015");
-const redux2i18n = require("./utils/redux2i18n");
 
 Promise.resolve(airTable.hydrateFromAirtable()).then(data => {
   // init i18next with serverside settings
@@ -45,18 +44,6 @@ Promise.resolve(airTable.hydrateFromAirtable()).then(data => {
           server.use(helmet());
           // enable middleware for i18next
           server.use(i18nextMiddleware.handle(i18nInstance));
-
-          // serve locales for client
-          server.use(
-            "/locales",
-            express.static(path.join(__dirname, "/locales"))
-          );
-
-          // missing keys
-          server.post(
-            "/locales/add/:lng/:ns",
-            i18nextMiddleware.missingKeyHandler(i18nInstance)
-          );
 
           // submitting Feedback
           server.post("/submitComment", (req, res) => {
@@ -112,28 +99,4 @@ Promise.resolve(airTable.hydrateFromAirtable()).then(data => {
         });
       }
     );
-  redux2i18n.redux2i18n(i18nInstance, data.translations);
 });
-
-// this code should run when the data-validation page is loaded and send that page the results
-// note that we can't check the urls in the browser because of CO
-//
-// var checkLinks = async function checklinks(benefits) {
-//   var brokenLinks = [];
-//   var responseEn, responseFr;
-//   for (let benefit of benefits) {
-//     responseEn = await fetch(benefit.benefitPageEn);
-//     responseFr = await fetch(benefit.benefitPageFr);
-//     if (responseEn.status !== 200) {
-//       brokenLinks.push(benefit.benefitPageEn);
-//     }
-//     if (responseFr.status !== 200) {
-//       brokenLinks.push(benefit.benefitPageFr);
-//     }
-//     console.log("type", typeof responseEn.body);
-//     if (responseEn.body.indexOf(benefit.benefitPageEn) === -1) {
-//       console.log(benefit.benefitPageEn, responseEn, "BAD URL");
-//     }
-//   }
-//   return brokenLinks;
-// };
